Updating redirect process to hide speedmarks id
This redirects to an html page at /redirect which has
a JavaScript redirect to the desired page. This prevents
the browser from sending the bookmarklet id to other
websites. For example, previously, if I clicked on a
link it would go to a site but they could look at the
referer header and get my bookmarklet id. With it, they
could delete all of my speedmarks. This is a bad thing.

Paul Fisher pointed this bug out quite some time ago.
I am fixing it now because I think he is a great guy.